+++ packages/driver-adapter-mysql2/CHANGELOG.md
@@ -1,8 +1,8 @@
-# @ashiba/driver-adapter-mysql2
+# @ashiba-ts/driver-adapter-mysql2
 
 ## 0.0.1
 
 ### Patch Changes
 
 - Updated dependencies [[`b451b4d`](https://github.com/mk3008/ashiba/commit/b451b4dcf395e71d2e68351880a013e9ba3a4546), [`d8e0689`](https://github.com/mk3008/ashiba/commit/d8e0689dd98d1e26eee6579c94113531307b8c2a), [`1c70b21`](https://github.com/mk3008/ashiba/commit/1c70b21d1ad826f28dc4965e1bfaa81a12771738), [`6521842`](https://github.com/mk3008/ashiba/commit/6521842760401bffee7cca6b48988a0d47f6e7f1)]:
-  - @ashiba/driver-adapter-core@0.1.0
+  - @ashiba-ts/driver-adapter-core@0.1.0